# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pharmacy API",
        "description": "RESTful API theo chuẩn JSON:API",
        "contact": {
            "email": "dongockhanh2003.utt@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Main Server"
        }
    ],
    "paths": {
...
        "/v1/reviews/write": {
            "post": {
                "tags": [
                    "Review"
                ],
                "summary": "Viết đánh giá",
                "description": "Viết đánh giá cho sản phẩm",
                "operationId": "writeReview",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "medicine_id",
                                    "rating",
                                    "comment"
                                ],
                                "properties": {
                                    "medicine_id": {
                                        "description": "ID của thuốc",
                                        "type": "string",
                                        "example": "1234567890"
                                    },
                                    "rating": {
                                        "description": "Đánh giá từ 1-5 sao",
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "comment": {
                                        "description": "Nội dung đánh giá",
                                        "type": "string",
                                        "example": "Sản phẩm rất tốt!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Đánh giá đã được thêm thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Đánh giá đã được thêm thành công"
                                        },
                                        "data": {
                                            "properties": {
                                                "_id": {
                                                    "type": "string",
                                                    "example": "507f1f77bcf86cd799439011"
                                                },
                                                "user_id": {
                                                    "type": "string",
                                                    "example": "507f1f77bcf86cd799439012"
                                                },
                                                "medicine_id": {
                                                    "type": "string",
                                                    "example": "507f1f77bcf86cd799439013"
                                                },
                                                "rating": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "comment": {
                                                    "type": "string",
                                                    "example": "Sản phẩm rất tốt!"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2023-01-01T00:00:00Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Dữ liệu không hợp lệ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Thông tin đánh giá không hợp lệ"
                                        },
                                        "data": {
                                            "type": "null",
                                            "example": null
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Chưa xác thực",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Không có quyền truy cập"
                                        },
                                        "data": {
                                            "type": "null",
                                            "example": null
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/reviews/update/{id}": {
            "patch": {
                "tags": [
                    "Review"
                ],
                "summary": "Cập nhật đánh giá",
                "description": "Cập nhật đánh giá sản phẩm (chỉ tác giả mới có quyền cập nhật)",
                "operationId": "updateReview",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID của đánh giá",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "medicine_id",
                                    "rating",
                                    "comment"
                                ],
                                "properties": {
                                    "medicine_id": {
                                        "description": "ID của thuốc",
                                        "type": "string",
                                        "example": "1234567890"
                                    },
                                    "rating": {
                                        "description": "Đánh giá từ 1-5 sao",
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "comment": {
                                        "description": "Nội dung đánh giá",
                                        "type": "string",
                                        "example": "Sản phẩm rất tốt!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cập nhật thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Đánh giá đã được cập nhật thành công"
                                        },
                                        "data": {
                                            "properties": {
                                                "_id": {
                                                    "type": "string",
                                                    "example": "507f1f77bcf86cd799439011"
                                                },
                                                "user_id": {
                                                    "type": "string",
                                                    "example": "507f1f77bcf86cd799439012"
                                                },
                                                "medicine_id": {
                                                    "type": "string",
                                                    "example": "507f1f77bcf86cd799439013"
                                                },
                                                "rating": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "comment": {
                                                    "type": "string",
                                                    "example": "Sản phẩm rất tốt!"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2023-01-01T00:00:00Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Không có quyền cập nhật",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Bạn không có quyền cập nhật đánh giá này"
                                        },
                                        "data": {
                                            "type": "null",
                                            "example": null
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Đánh giá không tồn tại",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Đánh giá không tồn tại"
                                        },
                                        "data": {
                                            "type": "null",
                                            "example": null
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Dữ liệu không hợp lệ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Thông tin đánh giá không hợp lệ"
                                        },
                                        "data": {
                                            "type": "null",
                                            "example": null
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Xác thực bằng JWT Token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}